Extract RecoveryClusterStateDelayListeners
This utility class is needed in another test in serverless, so this
commit extracts it to the test framework.
