+++ server/src/internalClusterTest/java/org/elasticsearch/indices/recovery/IndexRecoveryIT.java
@@ -39,7 +39,6 @@ import org.elasticsearch.action.index.IndexResponse;
 import org.elasticsearch.action.support.ActiveShardCount;
 import org.elasticsearch.action.support.ChannelActionListener;
 import org.elasticsearch.action.support.PlainActionFuture;
-import org.elasticsearch.action.support.SubscribableListener;
 import org.elasticsearch.action.support.WriteRequest.RefreshPolicy;
 import org.elasticsearch.action.support.replication.ReplicationResponse;
 import org.elasticsearch.cluster.ClusterState;
@@ -71,10 +70,6 @@ import org.elasticsearch.common.settings.Settings;
 import org.elasticsearch.common.unit.ByteSizeUnit;
 import org.elasticsearch.common.unit.ByteSizeValue;
 import org.elasticsearch.common.util.CollectionUtils;
-import org.elasticsearch.common.util.concurrent.ConcurrentCollections;
-import org.elasticsearch.core.AbstractRefCounted;
-import org.elasticsearch.core.RefCounted;
-import org.elasticsearch.core.Releasable;
 import org.elasticsearch.gateway.ReplicaShardAllocatorIT;
 import org.elasticsearch.index.Index;
 import org.elasticsearch.index.IndexService;
@@ -1643,54 +1638,7 @@ public class IndexRecoveryIT extends AbstractIndexRecoveryIntegTestCase {
 
         final long initialClusterStateVersion = clusterService().state().version();
 
-        // Helper class to encapsulate the sync mechanism that delays applying cluster states on the primary node until the replica gives
-        // the go-ahead.
-        class ClusterStateSyncListeners implements Releasable {
-            private final Map<Long, SubscribableListener<Void>> clusterStateBarriers = ConcurrentCollections.newConcurrentMap();
-            private final SubscribableListener<Void> startRecoveryListener = new SubscribableListener<>();
-
-            private final CountDownLatch completeLatch = new CountDownLatch(1);
-            private final RefCounted refCounted = AbstractRefCounted.of(completeLatch::countDown);
-            private final List<Runnable> cleanup = new ArrayList<>(2);
-
-            @Override
-            public void close() {
-                refCounted.decRef();
-                safeAwait(completeLatch);
-                cleanup.forEach(Runnable::run);
-                clusterStateBarriers.values().forEach(l -> l.onResponse(null));
-            }
-
-            void addCleanup(Runnable runnable) {
-                cleanup.add(runnable);
-            }
-
-            SubscribableListener<Void> getStateApplyDelayListener(long clusterStateVersion) {
-                assertThat(clusterStateVersion, greaterThanOrEqualTo(initialClusterStateVersion));
-                if (refCounted.tryIncRef()) {
-                    try {
-                        return clusterStateBarriers.computeIfAbsent(clusterStateVersion, ignored -> new SubscribableListener<>());
-                    } finally {
-                        refCounted.decRef();
-                    }
-                } else {
-                    return SubscribableListener.newSucceeded(null);
-                }
-            }
-
-            void onStartRecovery() {
-                Thread.yield();
-                assertFalse(startRecoveryListener.isDone());
-                startRecoveryListener.onResponse(null);
-            }
-
-            public void delayUntilRecoveryStart(SubscribableListener<Void> listener) {
-                assertFalse(startRecoveryListener.isDone());
-                startRecoveryListener.addListener(listener);
-            }
-        }
-
-        try (var clusterStateSyncListeners = new ClusterStateSyncListeners()) {
+        try (var recoveryClusterStateDelayListeners = new RecoveryClusterStateDelayListeners(initialClusterStateVersion)) {
             final var primaryNodeTransportService = (MockTransportService) internalCluster().getInstance(
                 TransportService.class,
                 primaryNode
@@ -1699,7 +1647,7 @@ public class IndexRecoveryIT extends AbstractIndexRecoveryIntegTestCase {
                 Coordinator.COMMIT_STATE_ACTION_NAME,
                 (handler, request, channel, task) -> {
                     assertThat(request, instanceOf(ApplyCommitRequest.class));
-                    clusterStateSyncListeners.getStateApplyDelayListener(((ApplyCommitRequest) request).getVersion())
+                    recoveryClusterStateDelayListeners.getClusterStateDelayListener(((ApplyCommitRequest) request).getVersion())
                         .addListener(
                             ActionListener.wrap(ignored -> handler.messageReceived(request, channel, task), e -> fail(e, "unexpected"))
                         );
@@ -1719,28 +1667,28 @@ public class IndexRecoveryIT extends AbstractIndexRecoveryIntegTestCase {
                         ),
                         task
                     );
-                    clusterStateSyncListeners.onStartRecovery();
+                    recoveryClusterStateDelayListeners.onStartRecovery();
                 }
             );
-            clusterStateSyncListeners.addCleanup(primaryNodeTransportService::clearInboundRules);
+            recoveryClusterStateDelayListeners.addCleanup(primaryNodeTransportService::clearInboundRules);
 
             final var replicaClusterService = internalCluster().getInstance(ClusterService.class, replicaNode);
             final ClusterStateListener clusterStateListener = event -> {
-                final var primaryProceedListener = clusterStateSyncListeners.getStateApplyDelayListener(event.state().version());
+                final var primaryProceedListener = recoveryClusterStateDelayListeners.getClusterStateDelayListener(event.state().version());
                 final var indexRoutingTable = event.state().routingTable().index(indexName);
                 assertNotNull(indexRoutingTable);
                 final var indexShardRoutingTable = indexRoutingTable.shard(0);
                 if (indexShardRoutingTable.size() == 2 && indexShardRoutingTable.getAllInitializingShards().isEmpty() == false) {
                     // this is the cluster state update which starts the recovery, so delay the primary node application until recovery
                     // has started
-                    clusterStateSyncListeners.delayUntilRecoveryStart(primaryProceedListener);
+                    recoveryClusterStateDelayListeners.delayUntilRecoveryStart(primaryProceedListener);
                 } else {
                     // this is some other cluster state update, so we must let it proceed now
                     primaryProceedListener.onResponse(null);
                 }
             };
             replicaClusterService.addListener(clusterStateListener);
-            clusterStateSyncListeners.addCleanup(() -> replicaClusterService.removeListener(clusterStateListener));
+            recoveryClusterStateDelayListeners.addCleanup(() -> replicaClusterService.removeListener(clusterStateListener));
 
             updateIndexSettings(Settings.builder().put(IndexMetadata.SETTING_NUMBER_OF_REPLICAS, 1), indexName);
             ensureGreen(indexName);
